+++ psutil/__init__.py
@@ -681,7 +681,7 @@ class Process(object):
             """
             if ioclass is None:
                 if value is not None:
-                    raise ValueError("'ioclass' must be specified")
+                    raise ValueError("'ioclass' argument must be specified")
                 return self._proc.ionice_get()
             else:
                 return self._proc.ionice_set(ioclass, value)
@@ -1007,10 +1007,12 @@ class Process(object):
 
     if _POSIX:
         def _send_signal(self, sig):
-            # XXX: according to "man 2 kill" PID 0 has a special
-            # meaning as it refers to <<every process in the process
-            # group of the calling process>>, so should we prevent
-            # it here?
+            if self.pid == 0:
+                # see "man 2 kill"
+                raise ValueError(
+                    "preventing sending signal to process with PID 0 as it "
+                    "would affect every process in the process group of the "
+                    "calling process (os.getpid()) instead of PID 0")
             try:
                 os.kill(self.pid, sig)
             except OSError as err:
@@ -1853,14 +1855,6 @@ def test():
                                     time.localtime(sum(pinfo['cpu_times'])))
             try:
                 user = p.username()
-            except KeyError:
-                if _POSIX:
-                    if pinfo['uids']:
-                        user = str(pinfo['uids'].real)
-                    else:
-                        user = ''
-                else:
-                    raise
             except Error:
                 user = ''
             if _WINDOWS and '\\' in user: